pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aa-structuretimers"
dynamic = ["version"]
description = "An app for keeping track of Eve Online structure timers with Alliance Auth and Discord"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Erik Kalkoken", email = "kalkoken87@gmail.com" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth-app-utils>=1.17.1",
    "allianceauth>=3",
    "dhooks_lite>=1.0.0",
    "django-eveuniverse>=0.16",
    "django-multiselectfield",
    "redis-simple-mq>=0.4",
    "requests",
]

[project.urls]
Homepage = "https://gitlab.com/ErikKalkoken/aa-structuretimers"
Source = "https://gitlab.com/ErikKalkoken/aa-structuretimers"
Changelog = "https://gitlab.com/ErikKalkoken/aa-structuretimers/-/blob/master/CHANGELOG.md"
Tracker = "https://gitlab.com/ErikKalkoken/aa-structuretimers/-/issues"

[tool.hatch.version]
path = "structuretimers/__init__.py"

[tool.hatch.build]
include = [
    "/structuretimers",
]

[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eveuniverse"]

[tool.pylint.'MASTER']
ignore-patterns = ["test_.*.py,__init__.py", "generate_.*.py"]

[tool.pylint.'BASIC']
good-names = ["i", "j", "k", "x", "f", "ex"]

[tool.pylint.'FORMAT']
max-line-length = 100

[tool.pylint.'MESSAGES CONTROL']
disable = ["R", "C"]
